+++ os-sanitizer-ebpf/src/vfs_fstatat.rs
@@ -23,7 +23,7 @@ unsafe fn try_fentry_vfs_fstatat(ctx: &FEntryContext) -> Result<u32, OsSanitizer
     let usermode_ptr: uintptr_t = ctx.arg(1);
 
     ACCESS_MAP
-        .insert(&(pid_tgid, dfd, usermode_ptr), &Stat, 0)
+        .insert(&(pid_tgid, dfd as u64, usermode_ptr as u64), &Stat, 0)
         .map_err(|_| Unreachable("map insertion failure"))?;
 
     Ok(0)
